pyproject.toml
[project]
name = "agnoagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.1.1",
    "duckduckgo-search>=8.1.1",
    "lancedb>=0.25.1",
    "langchain-core>=0.3.78",
    "pandas>=2.3.3",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "tantivy>=0.25.0",
    "yfinance>=0.2.66",
]
